Use MultiWriterIDGenerator for receipts
